+++ server/manifest.go
@@ -12,10 +12,10 @@ var manifest *model.Manifest
 
 const manifestStr = `
 {
-  "id": "com.mattermost.mattermost-teamcity-plugin",
+  "id": "mattermost-teamcity-plugin",
   "name": "Mattermost TeamCity Plugin",
   "description": "This plugin integrates Mattermost with TeamCity",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "min_server_version": "5.18.0",
   "server": {
     "executables": {