+++ webapp/src/manifest.js
@@ -2,10 +2,10 @@
 
 const manifest = JSON.parse(`
 {
-    "id": "com.mattermost.mattermost-teamcity-plugin",
+    "id": "mattermost-teamcity-plugin",
     "name": "Mattermost TeamCity Plugin",
     "description": "This plugin integrates Mattermost with TeamCity",
-    "version": "0.1.0",
+    "version": "0.2.0",
     "min_server_version": "5.18.0",
     "server": {
         "executables": {