+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 0.0.1 - 2018-08-16
+## Unreleased
+- `/teamcity stats` to return statistics
+- Incoming Webhook to handle TeamCity Events
+
+## 0.1.0 - 2020-01-19
 ### Added
-- Initial release
+- `/teamcity enable`
+- `/teamcity disable`
+- `/teamcity install`
+- `/teamcity list projects`
+- `/teamcity list builds`
+- `/teamcity build start`
+- `/teamcity build cancel`